fix(loss): admit fp16 logits in the chunked log-prob path

+++ miles/backends/training_utils/loss_hub/losses.py
@@ -333,9 +333,9 @@ def policy_loss_function(
         if args.kl_loss_coef != 0:
             loss = loss + args.kl_loss_coef * kl_loss
 
-    # make sure the gradient could backprop correctly.
+    # make sure the gradient could backprop correctly; fp32 sum avoids fp16 inf -> nan
     if log_probs.numel() == 0:
-        loss += 0 * logits.sum()
+        loss += 0 * logits.sum(dtype=torch.float32)
 
     train_scored_log_probs = old_log_probs
     train_rollout_logprob_abs_diff = None
